Cell shutdown process has been fixed yet again
Previously, the cell would just shut down and not even care about
the state of bands and stuff. Now, all portions of the cell have
stopNotify channels so the shutdown process can wait for each thing
to stop before it shuts down completely. In addition to this, the
shutdown process of scribe is now awaited, so the entire process
will get logged no matter what.